test #66: add validation test
