+++ API-Server/src/test/java/swm/hkcc/LGTM/app/modules/registration/service/RegistrationServiceTest.java
@@ -10,6 +10,7 @@ import swm.hkcc.LGTM.app.modules.member.domain.Member;
 import swm.hkcc.LGTM.app.modules.member.domain.Senior;
 import swm.hkcc.LGTM.app.modules.member.exception.NotJuniorMember;
 import swm.hkcc.LGTM.app.modules.member.exception.NotSeniorMember;
+import swm.hkcc.LGTM.app.modules.member.service.MemberValidator;
 import swm.hkcc.LGTM.app.modules.mission.domain.Mission;
 import swm.hkcc.LGTM.app.modules.mission.exception.NotExistMission;
 import swm.hkcc.LGTM.app.modules.mission.repository.MissionRepository;
@@ -32,25 +33,18 @@ class RegistrationServiceTest {
     @Mock
     private MissionRegistrationRepository missionRegistrationRepository;
 
+    @Mock
+    private MemberValidator memberValidator;
+
+    @Mock
+    private RegistrationValidator registrationValidator;
+
     @InjectMocks
     private RegistrationService registrationService;
 
     @BeforeEach
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-    }
-
-    @Test
-    void Register_junior가_아닌_경우() {
-        // given
-        Member notJunior = Member.builder()
-                .memberId(1L)
-                .build();
-        // when
-
-        // then
-        assertThatThrownBy(() -> registrationService.registerJunior(notJunior, 1L))
-                .isInstanceOf(NotJuniorMember.class);
     }
 
     @Test
@@ -70,68 +64,6 @@ class RegistrationServiceTest {
                 .isInstanceOf(NotExistMission.class);
     }
 
-    // 미션 마감일이 지난 경우
-    @Test
-    void Register_미션_마감일이_지난_경우() {
-        // given
-        Member junior = Member.builder()
-                .memberId(1L)
-                .junior(Junior.builder()
-                        .juniorId(1L)
-                        .build())
-                .build();
-        Mission mission = Mission.builder()
-                .missionId(1L)
-                .registrationDueDate(LocalDate.now().minusDays(1))
-                .build();
-        given(missionRepository.findById(1L)).willReturn(Optional.of(mission));
-        given(missionRegistrationRepository.countByMission_MissionId(1L)).willReturn(10);
-
-        // when
-
-        // then
-        assertThatThrownBy(() -> registrationService.registerJunior(junior, mission.getMissionId()))
-                .isInstanceOf(MissRegisterDeadline.class);
-    }
-
-    @Test
-    void Register_미션_등록인원이_넘치는_경우() {
-        // given
-        Member junior = Member.builder()
-                .memberId(1L)
-                .junior(Junior.builder()
-                        .juniorId(1L)
-                        .build())
-                .build();
-        Mission mission = Mission.builder()
-                .missionId(1L)
-                .registrationDueDate(LocalDate.now().plusDays(1))
-                .maxPeopleNumber(10)
-                .build();
-        given(missionRepository.findById(1L)).willReturn(Optional.of(mission));
-        given(missionRegistrationRepository.countByMission_MissionId(1L)).willReturn(10);
-
-
-        // when
-
-        // then
-        assertThatThrownBy(() -> registrationService.registerJunior(junior, mission.getMissionId()))
-                .isInstanceOf(FullRegisterMembers.class);
-    }
-
-    @Test
-    void Senior_Info_Senior가_아닌_경우() {
-        // given
-        Member notSenior = Member.builder()
-                .memberId(1L)
-                .build();
-        // when
-
-        // then
-        assertThatThrownBy(() -> registrationService.getSeniorEnrollInfo(notSenior, 1L))
-                .isInstanceOf(NotSeniorMember.class);
-    }
-
     @Test
     void Senior_Info_미션_없는_경우() {
         // given
@@ -149,28 +81,4 @@ class RegistrationServiceTest {
                 .isInstanceOf(NotExistMission.class);
     }
 
-    @Test
-    void Senior_Info_자신의_미션이_아닌_경우() {
-        // given
-        Member senior = Member.builder()
-                .memberId(1L)
-                .senior(Senior.builder()
-                        .seniorId(1L)
-                        .build())
-                .build();
-        Mission mission = Mission.builder()
-                .missionId(1L)
-                .writer(Member.builder()
-                        .memberId(2L)
-                        .build())
-                .registrationDueDate(LocalDate.now().plusDays(1))
-                .maxPeopleNumber(10)
-                .build();
-
-        given(missionRepository.findById(1L)).willReturn(Optional.ofNullable(mission));
-
-
-        assertThatThrownBy(() -> registrationService.getSeniorEnrollInfo(senior, 1L))
-                .isInstanceOf(NotMyMission.class);
-    }
 }
